[ALPS03585046] cirq: fix coverity issue
[Detail]
fix "out-of-bounds access" problem

MTK-Commit-Id: b54ab46e57e8cac0a935dc3e242026a27c69e8eb

CR-Id: ALPS03585046
Feature: cirq

+++ drivers/misc/mediatek/cirq/mtk_sys_cirq.h
@@ -94,7 +94,7 @@ static void cirq_fast_sw_flush(void);
 struct cirq_reg {
 	unsigned int reg_num;
 	unsigned int used;
-	unsigned int mask;
+	unsigned long mask;
 	unsigned int pol;
 	unsigned int sen;
 	unsigned long pending;
